reorganize code and better error handling
