+++ crates/rnote-ui/src/workspacebrowser/workspaceactions/mod.rs
@@ -1,6 +1,7 @@
 // Modules
 mod createfolder;
+mod openfolder;
 
 // Re-exports
 pub(crate) use createfolder::create_folder;
-pub(crate) use createfolder::open_folder;
+pub(crate) use openfolder::open_folder;
